Refactor LeaderboardCacheService
Saves are now saved with out the position. This is needed for the more complex filter system, because we don't wanna save each individual filter option inside the redis cache and keeping the pre calculated positions only for the raw data makes no sense, because this is the data least requested.

- Remove filter option
- Remove position from cache